refactor(auth): extract Google sign-in logic from NextAuth config
Moves handleGoogleSignIn and splitGoogleName into service/auth/signup.ts
so the Google account-creation flow is testable without a full OAuth
round-trip. The NextAuth signIn callback becomes pure wiring: it resolves
preconditions (googleId, email) then delegates.
